Rename setup_profile -> setup_nanopub_profile and make test work
